+++ Matrix/TPZPardisoControl.h
@@ -81,39 +81,39 @@ protected:
     MProperty fProperty;
     
     // Solver internal data address pointers
-    // 32-bit: int pt[64]; 64-bit: long int pt[64]
+    // 32-bit: int pt[64]; 64-bit: int64_t pt[64]
     // or void *pt[64] should be OK on both architectures
     // this datastructure should not be copied or duplicated, therefore the "autopointer" protection
     
-    TPZAutoPointer<TPZManVector<long long, 64> > fPardisoControl;
+    TPZAutoPointer<TPZManVector<int64_t, 64> > fPardisoControl;
     
     // adress of the first element of pt;
-    long long *fHandle;
-    //  ConcreteRigidArray1d<long int, 64> pt;
+    int64_t *fHandle;
+    //  ConcreteRigidArray1d<int64_t, 64> pt;
     
     // Array used to pass parameters to Pardiso
-    TPZManVector<long long, 64> fParam;
+    TPZManVector<int64_t, 64> fParam;
     
     // Maximum number of factors we will pass to the solver
-    long long fMax_num_factors;
+    int64_t fMax_num_factors;
     
     // Factor number we are using
-    long long fMatrix_num;
+    int64_t fMatrix_num;
     
     // Message level information
-    long long fMessageLevel;
+    int64_t fMessageLevel;
     
     // error flag from Pardiso
-    long long fError;
+    int64_t fError;
     
     /// permutation vector computed by Pardiso
-    TPZVec<long long> fPermutation;
+    TPZVec<int64_t> fPermutation;
     
     // matrix type, computed based on the structural information and TVar
-    long long fMatrixType;
+    int64_t fMatrixType;
     
     /// Compute the matrix type
-    long long MatrixType();
+    int64_t MatrixType();
     
     /// pointer to the nonsymmetric system (where the data structures are stored
     TPZFYsmpMatrix<TVar> *fNonSymmetricSystem;
